chore: move TechStackBadge file
